moved utlshandler into netxlite
